requirements.txt
-i https://pypi.org/simple
alembic==1.14.1; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
boto3==1.36.10; python_version >= '3.8'
botocore==1.36.10; python_version >= '3.8'
certifi==2024.12.14; python_version >= '3.6'
click>=8; python_version >= '2.7'
cloudinary==1.42.1
flask==2.2.5; python_version >= '2.7'
flask-admin==1.6.1; python_version >= '3.6'
flask-basicauth==0.2.0
flask-cors==5.0.0
flask-jwt-extended==4.7.1; python_version >= '3.9' and python_version < '4'
pyjwt>=2.0,<3.0; python_version >= '3.9'
gevent==24.11.1; python_version >= '3.9'
greenlet==3.1.1; platform_python_implementation == 'CPython'
gunicorn==23.0.0; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
mako==1.3.8; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
s3transfer==0.11.2; python_version >= '3.8'
setuptools==75.8.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlalchemy==2.0.37; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.3.0; python_version >= '3.9'
wtforms==3.2.1; python_version >= '3.9'
zope.event==5.0; python_version >= '3.7'
zope.interface==7.2; python_version >= '3.8'
flask-migrate==4.0.5
flask-swagger==0.2.14
Flask-Mail==0.9.1

